+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 3. add `SupportMotion` loads, including `SupportDisplacement`, `SupportVelocity` and `SupportAcceleration`
 4. update `Armadillo` to version 10.4
 5. add functionality to check new version
+6. add `FEAST` solver
 
 ## version 1.3